Se añadió en el store de pinia un archivo global en el que se configuró el state en el que se grabara el activeUsername (de capital importancia para vincularlo con quien es el autor de la receta a los fines de que sea este el único que cuente con los permisos suficientes para modificar u elmnar la misma), dos actions una para establecer el activeUsername y otra ara el logout(se añadió para esto un componente de logout para cerrar la sesión que invoca a un metodo en global), y dos getters uno para obtener el nombre del usuario y otro para saber si esta logeado (se usara este ult por ej para q el navbar muestre lo de añadir receta /lista de receta solo si el usuario esta logeado, e inversamente si no lo esta se le mostrara q inicie sesion o se registre)  Queda mejorar la presentacion con bootstrap

+++ mi-proyecto/src/components/ListaDeRecetas.vue
@@ -11,7 +11,7 @@
         <img v-if="receta.imagen" :src="receta.imagen" alt="Imagen de la receta" class="img-fluid rounded" style="max-height: 100px;">
         <div>
           <button @click="eliminarReceta(receta.id)" class="btn btn-danger rounded">Eliminar</button>
-          <button @click="irAModificarReceta(receta.id)" class="btn btn-danger rounded">Modificar uwu</button>
+          <button @click="irAModificarReceta(receta.id)" class="btn btn-danger rounded">Modificar</button>
 
         </div>
 
@@ -23,11 +23,13 @@
 
 <script>
 import axios from 'axios';
+import {useGlobalStore} from "../../store/global.js";
 
 export default {
   data() {
     return {
       recetas: [],
+      globalStore: useGlobalStore(),
     };
   },
   created() {
@@ -44,16 +46,28 @@ export default {
       }
     },
     async eliminarReceta(id) {
-      try {
-        const response = await axios.delete(`https://670ed6f63e7151861655ee25.mockapi.io/uwu/recetas/${id}`);
-        this.recetas = this.recetas.filter(receta => receta.id !== id);
-        alert('Pude eliminar la receta!')
-      } catch (error) {
-        console.error('Error al eliminar la receta:', error);
+      const receta = this.recetas.find(r=>r.id===id);
+      if (receta.autor === this.globalStore.getActiveUsername) {
+        try {
+          const response = await axios.delete(`https://670ed6f63e7151861655ee25.mockapi.io/uwu/recetas/${id}`);
+          this.recetas = this.recetas.filter(receta => receta.id !== id);
+          alert('Pude eliminar la receta!')
+        } catch (error) {
+          console.error('Error al eliminar la receta:', error);
+        }
+      }
+      else{
+        alert('No tenes los permisos suficientes para eliminar esta receta xq no sos su autor')
       }
     },
     irAModificarReceta(id) {
-      this.$router.push(`/modificar/${id}`);
+      const receta = this.recetas.find(r=>r.id===id);
+      if (receta.autor === this.globalStore.getActiveUsername) {
+        this.$router.push(`/modificar/${id}`);
+      }
+      else{
+        alert('No posees los permisos para modificar esta receta ya q no sos su autor')
+      }
     }
